final commit

+++ src/main/java/com/stacksimplify/restservices/entities/Employee.java
@@ -14,17 +14,15 @@ public class Employee {
 	private Long empid;
 	private String fname;
 	private String lname;
-	private Integer age;
+	private Long age;
 	private String designation;
 	private String address;
-	
 	
 	public Employee() {
 		
 	}
 
-
-	public Employee(Long empid, String fname, String lname, Integer age, String designation, String address) {
+	public Employee(Long empid, String fname, String lname, Long age, String designation, String address) {
 		super();
 		this.empid = empid;
 		this.fname = fname;
@@ -34,66 +32,53 @@ public class Employee {
 		this.address = address;
 	}
 
-
 	public Long getEmpid() {
 		return empid;
 	}
-
 
 	public void setEmpid(Long empid) {
 		this.empid = empid;
 	}
 
-
 	public String getFname() {
 		return fname;
 	}
-
 
 	public void setFname(String fname) {
 		this.fname = fname;
 	}
 
-
 	public String getLname() {
 		return lname;
 	}
-
 
 	public void setLname(String lname) {
 		this.lname = lname;
 	}
 
-
-	public Integer getAge() {
+	public Long getAge() {
 		return age;
 	}
 
-
-	public void setAge(Integer age) {
+	public void setAge(Long age) {
 		this.age = age;
 	}
-
 
 	public String getDesignation() {
 		return designation;
 	}
 
-
 	public void setDesignation(String designation) {
 		this.designation = designation;
 	}
-
 
 	public String getAddress() {
 		return address;
 	}
 
-
 	public void setAddress(String address) {
 		this.address = address;
 	}
-
 
 	@Override
 	public String toString() {
@@ -101,8 +86,6 @@ public class Employee {
 				+ designation + ", address=" + address + "]";
 	}
 	
-	
-	
-	
+		
 
 }
